removed old StationInitializer :fire:
