--- resources/coordsMap.txt
@@ -1,200 +1,0 @@
-0   0           0
-1	318			78
-2	782			42
-3	1082		48
-4	1268		36
-5	1996		58
-6	2224		58
-7	2444		74
-8	212			186
-9	402			190
-10	934			176
-11	1072		200
-12	1186		176
-13	1398		170
-14	1614		124
-15	1832		104
-16	2040		182
-17	2428		250
-18	110			268
-19	288			290
-20	500			236
-21	724			326
-22	1082		360
-23	1252		270
-24	1522		282
-25	1640		310
-26	1814		178
-27	1846		282
-28	1924		242
-29	2220		304
-30	2506		290
-31	188			350
-32	440			410
-33	626			378
-34	948			410
-35	1134		458
-36	1212		468
-37	1326		368
-38	1578		376
-39	1692		358
-40	1892		440
-41	1976		404
-42	2434		406
-43	68			444
-44	320			496
-45	506			526
-46	660			480
-47	782			444
-48	982			536
-49	1278		542
-50	1400		464
-51	1648		488
-52	1774		456
-53	1912		534
-54	2008		502
-55	2228		498
-56	2516		526
-57	152			542
-58	396			570
-59	442			628
-60	538			614
-61	706			638
-62	792			610
-63	1004		716
-64	1132		690
-65	1264		666
-66	1344		648
-67	1492		622
-68	1666		586
-69	1826		574
-70	2016		606
-71	2198		608
-72	2384		628
-73	150			642
-74	232			740
-75	356			704
-76	510			704
-77	610			790
-78	742			766
-79	832			750
-80	1038		792
-81	1224		820
-82	1294		776
-83	1460		746
-84	1578		696
-85	1676		664
-86	1848		728
-87	2030		776
-88	2116		792
-89	2186		730
-90	2308		734
-91	2490		740
-92	84			834
-93	96			900
-94	250			870
-95	328			854
-96	708			900
-97	772			880
-98	878			842
-99	988			854
-100	1158		912
-101	1332		848
-102	1558		762
-103	1704		742
-104	1850		812
-105	2256		846
-106	2410		868
-107	2526		872
-108	2218		1016
-109	814			1034
-110	940			918
-111	1016		1002
-112	1060		975
-113	1232		986
-114	1356		942
-115	1556		882
-116	1852		984
-117	2058		1060
-118	1856		1104
-119	2474		1142
-120	80			1218
-121	168			1222
-122	292			1214
-123	610			1208
-124	788			1164
-125	1118		1058
-126	1440		1004
-127	1676		1060
-128	1974		1406
-129	2042		1128
-130	1062		1178
-131	1160		1112
-132	1362		1108
-133	1582		1214
-134	1744		1162
-135	2118		1200
-136	2412		1312
-137	498			1310
-138	844			1240
-139	1048		1254
-140	1356		1228
-141	1642		1252
-142	1852		1294
-143	2042		1268
-144	110			1438
-145	198			1428
-146	320			1420
-147	410			1398
-148	538			1378
-149	644			1364
-150	764			1318
-151	824			1382
-152	912			1320
-153	958			1392
-154	1172		1346
-155	1244		1438
-156	1386		1438
-157	1510		1448
-158	1712		1366
-159	1720		1644
-160	2128		1440
-161	2294		1422
-162	2530		1424
-163	308			1486
-164	436			1486
-165	674			1524
-166	924			1470
-167	1122		1508
-168	1192		1560
-169	1382		1536
-170	1492		1552
-171	2244		1820
-172	1910		1542
-173	2170		1620
-174	2388		1556
-175	2504		1648
-176	72			1612
-177	176			1588
-178	364			1572
-179	582			1596
-180	720			1624
-181	864			1586
-182	930			1604
-183	1062		1546
-184	1310		1646
-185	1450		1746
-186	1610		1712
-187	1842		1658
-188	2074		1662
-189	180			1758
-190	290			1824
-191	438			1706
-192	472			1858
-193	818			1718
-194	860			1768
-195	966			1760
-196	1116		1680
-197	1138		1784
-198	1714		1872
-199	2102		1858
