--- resources/neighbours.txt
@@ -1,275 +1,0 @@
-1	t	9 8 
-1	b	46 58 
-1	u	46 
-2	t	20 10 
-3	t	11 4 12 
-3	b	23 22 
-4	t	3 13 
-5	t	15 16 
-6	t	29 7 
-7	t	6 17 
-7	b	42 
-8	t	1 18 19 
-9	t	1 19 20 
-10	t	2 11 21 34 
-11	t	3 10 22 
-12	t	3 23 
-13	t	4 23 24
-13	b	23 14 52 
-13	u	46 67 89 
-14	t	13 15 25 
-14	b	13 15 
-15	t	5 14 16 26 28 
-15	b	14 41 29 
-16	t	5 15 28 29 
-17	t	7 29 30 
-18	t	8 31 43 
-19	t	8 9 32 
-20	t	2 9 33 
-21	t	10 33 
-22	t	11 23 34 35 
-22	b	3 23 34 65 
-23	t	12 13 22 37 
-23	b	3 13 22 67 
-24	t	37 38 
-25	t	14 38 39 
-26	t	15 27 39 
-27	t	26 28 40 
-28	t	15 16 27 41 
-29	t	6 16 17 41 42 
-29	b	15 41 42 55 
-30	t	17 42 
-31	t	18 43 44 
-32	t	19 33 44 45 
-33	t	20 21 32 46 
-34	t	10 22 47 48 
-34	b	22 46 63 
-35	t	22 36 48 65 
-36	t	35 37 49 
-37	t	23 24 36 50 
-38	t	24 25 50 51 
-39	t	25 26 51 52 
-40	t	27 41 52 53 
-41	t	28 29 40 54 
-41	b	15 29 52 87 
-42	t	29 30 56 72 
-42	b	7 29 72 
-43	t	18 31 57 
-44	t	31 32 58 
-45	t	32 46 58 59 60 
-46	t	33 45 47 61 
-46	b	1 34 58 78 
-46	u	1 13 74 79 
-47	t	34 46 62 
-48	t	34 35 62 63 
-49	t	36 50 66 
-50	t	37 38 49
-51	t	38 39 52 67 68 
-52	t	39 40 51 69 
-52	b	13 41 67 86 
-53	t	40 54 69 
-54	t	41 53 55 70 
-55	t	54 71 
-55	b	29 89 
-56	t	42 91 
-57	t	43 58 73 
-58	t	44 45 57 59 74 
-58	b	1 46 74 77 
-59	t	45 58 75 76 
-60	t	45 61 76 
-61	t	46 60 62 76 78 
-62	t	47 48 61 79 
-63	t	48 64 79 80 
-63	b	34 65 79 100 
-64	t	63 65 81 
-65	t	35 64 66 82 
-65	b	22 63 67 82 
-66	t	49 65 67 82 
-67	t	51 66 68 84
-67	b	23 52 65 82 102 
-67	u	13 79 89 111 
-68	t	51 67 69 85 
-69	t	52 53 68 86 
-70	t	54 71 87 
-71	t	55 70 72 89 
-72	t	42 71 90 91 
-72	b	42 105 107 
-73	t	57 74 92 
-74	t	58 73 75 92 
-74	b	58 94 
-74	u	46 
-75	t	59 74 94 
-76	t	59 60 61 77 
-77	t	76 78 95 96 
-77	b	58 78 94 124 
-78	t	61 77 79 97 
-78	b	46 77 79 
-79	t	62 63 78 98 
-79	b	63 78 
-79	u	46 67 93 111 
-80	t	63 99 100 
-81	t	64 82 100 
-82	t	65 66 81 101 
-82	b	65 67 100 140 
-83	t	101 102 
-84	t	67 85 
-85	t	68 84 103 
-86	t	69 103 104 
-86	b	52 87 102 116 
-87	t	70 88 
-87	b	41 86 105 
-88	t	87 89 117 
-89	t	71 88 105 
-89	b	55 105 
-89	u	13 67 128 140 
-90	t	72 91 105 
-91	t	56 72 90 105 107 
-92	t	73 74 93 
-93	t	92 94 
-93	b	94 
-93	u	79 
-94	t	75 93 95 
-94	b	74 77 93 
-95	t	77 94 122 
-96	t	77 97 109 
-97	t	78 96 98 109 
-98	t	79 97 99 110 
-99	t	80 98 110 112 
-100	t	80 81 101 112 113 
-100	b	63 82 111 
-101	t	82 83 100 114 
-102	t	83 103 115 
-102	b	67 86 127 
-103	t	85 86 102 
-104	t	86 116 
-105	t	89 90 91 106 108 
-105	b	72 87 89 107 108 
-106	t	105 107 
-107	t	91 106 119 
-107	b	72 105 161 
-108	t	105 117 119 
-108	b	105 116 135 
-109	t	96 97 110 124 
-110	t	98 99 109 111 
-111	t	110 112 124 
-111	b	100 124 
-111	u	67 79 153 163 
-112	t	99 100 111 125 
-113	t	100 114 125 
-114	t	101 113 115 126 131 132 
-115	t	102 114 126 127 
-116	t	104 117 118 127 
-116	b	86 108 127 142 
-117	t	88 108 116 129 
-118	t	116 129 134 142 
-119	t	107 108 136 
-120	t	121 144 
-121	t	120 122 145 
-122	t	95 121 123 146 
-122	b	123 144 
-123	t	122 124 137 148 149 
-123	b	122 124 144 165 
-124	t	109 111 123 130 138 
-124	b	77 111 123 153 
-125	t	112 113 131 
-126	t	114 115 127 140 
-127	t	115 116 126 133 134 
-127	b	102 116 133 
-128	t	142 143 160 172 188 
-128	b	135 142 161 187 199 
-128	u	89 140 185 
-129	t	117 118 135 142 143 
-130	t	124 131 139 
-131	t	114 125 130 
-132	t	114 140 
-133	t	127 140 141 
-133	b	127 140 157 
-134	t	118 127 141 142 
-135	t	129 136 143 161 
-135	b	108 128 161 
-136	t	119 135 162 
-137	t	123 147 
-138	t	124 150 152 
-139	t	130 140 153 154 
-140	t	126 132 133 139 154 156 
-140	b	82 133 154 156 
-140	u	89 128 153 
-141	t	133 134 142 158 
-142	t	118 128 129 134 141 143 158 
-142	b	116 128 157
-143	t	128 129 135 142 160 
-144	t	120 145 177 
-144	b	122 123 163 
-145	t	121 144 146 
-146	t	122 145 147 163 
-147	t	137 146 164 
-148	t	123 149 164 
-149	t	123 148 150 165 
-150	t	138 149 151 
-151	t	150 152 165 166 
-152	t	138 151 153 
-153	t	139 152 154 166 167 
-153	b	124 154 180 184 
-153	u	111 140 163 185 
-154	t	139 140 153 155 
-154	b	140 153 156 
-155	t	154 156 167 168 
-156	t	140 155 157 169 
-156	b	140 154 157 184 
-157	t	156 158 170 
-157	b	133 156 185 142
-158	t	141 142 157 159 
-159	t	158 170 172 186 198 
-160	t	128 143 161 173 
-161	t	135 160 174 
-161	b	107 128 135 199 
-162	t	136 175 
-163	t	146 177 
-163	b	144 176 191 
-163	u	111 153 
-164	t	147 148 178 179 
-165	t	149 151 179 180 
-165	b	123 180 191 
-166	t	151 153 181 183 
-167	t	153 155 168 183 
-168	t	155 167 184 
-169	t	156 184 
-170	t	157 159 185 
-171	t	173 175 199 
-172	t	128 159 187 
-173	t	160 171 174 188 
-174	t	161 173 175 
-175	t	162 171 174 
-176	t	177 189 
-176	b	163 190 
-177	t	144 163 176 
-178	t	164 189 191 
-179	t	164 165 191 
-180	t	165 181 193 
-180	b	153 165 184 190 
-181	t	166 180 182 193 
-182	t	181 183 195 
-183	t	166 167 182 196 
-184	t	168 169 185 196 197 
-184	b	153 156 180 185 
-185	t	170 184 186 
-185	b	157 184 187 
-185	u	128 153 
-186	t	159 185 198 
-187	t	172 188 198 
-187	b	128 185 
-188	t	128 173 187 199 
-189	t	176 178 190 
-190	t	189 191 192 
-190	b	176 180 191 
-191	t	178 179 190 192 
-191	b	163 165 190 
-192	t	190 191 194 
-193	t	180 181 194 
-194	t	192 193 195 
-195	t	182 194 197 
-196	t	183 184 197 
-197	t	184 195 196 
-198	t	159 186 187 199 
-199	t	171 188 198 
-199	b	128 161
