--- resources/tuiCoords.txt
@@ -1,200 +1,0 @@
-5: 214 5
-1: 46 6
-3: 123 6
-2: 81 7
-6: 238 8
-4: 140 11
-7: 259 11
-8: 31 13
-15: 194 13
-9: 53 14
-10: 97 14
-14: 171 15
-16: 219 16
-11: 114 17
-13: 154 18
-20: 66 19
-12: 130 19
-18: 18 20
-19: 38 21
-26: 191 23
-21: 84 24
-23: 136 26
-17: 256 27
-31: 26 28
-33: 72 30
-22: 115 30
-29: 236 30
-28: 207 31
-43: 9 32
-30: 269 32
-32: 49 33
-34: 99 33
-24: 157 35
-25: 173 37
-27: 196 37
-47: 89 39
-46: 76 41
-41: 214 41
-44: 34 42
-35: 119 43
-45: 56 44
-37: 145 44
-38: 163 45
-55: 236 45
-39: 181 46
-42: 260 46
-57: 16 47
-40: 198 48
-54: 214 49
-58: 41 51
-36: 131 52
-52: 187 53
-48: 101 56
-51: 170 56
-50: 151 57
-53: 203 57
-73: 12 59
-60: 64 59
-62: 89 59
-56: 276 60
-59: 51 61
-49: 137 61
-70: 214 61
-71: 235 61
-61: 79 62
-69: 193 63
-72: 260 64
-68: 179 66
-75: 41 68
-67: 157 69
-76: 58 71
-66: 141 71
-65: 132 72
-74: 24 73
-85: 178 73
-64: 118 74
-63: 104 76
-84: 164 76
-92: 5 78
-89: 233 78
-90: 253 78
-79: 93 79
-86: 193 79
-91: 270 79
-78: 82 80
-83: 149 80
-77: 67 81
-87: 210 82
-82: 135 83
-80: 109 84
-102: 159 84
-103: 178 84
-95: 36 85
-94: 27 87
-81: 126 87
-104: 193 88
-93: 9 89
-98: 97 90
-101: 138 90
-88: 226 90
-105: 246 91
-99: 109 93
-106: 261 93
-107: 275 93
-97: 83 94
-115: 156 95
-96: 73 97
-110: 101 99
-100: 121 101
-114: 141 104
-116: 194 105
-112: 112 106
-117: 224 106
-108: 245 106
-113: 127 109
-119: 271 110
-111: 106 112
-127: 177 113
-109: 84 115
-118: 194 115
-129: 220 115
-125: 118 118
-126: 148 119
-134: 187 122
-131: 128 125
-132: 140 125
-135: 232 125
-136: 271 128
-130: 111 130
-124: 80 131
-120: 4 135
-121: 15 135
-122: 31 135
-123: 63 135
-142: 198 135
-143: 219 136
-139: 111 139
-133: 170 140
-141: 184 140
-140: 141 141
-138: 91 144
-137: 57 147
-162: 276 147
-128: 211 151
-161: 256 151
-158: 187 153
-150: 83 154
-160: 229 154
-152: 103 158
-154: 132 158
-148: 61 159
-147: 43 161
-149: 71 161
-174: 264 162
-144: 6 164
-145: 19 164
-146: 34 164
-151: 93 165
-157: 171 165
-156: 150 166
-172: 208 168
-175: 276 168
-153: 116 175
-155: 136 175
-163: 31 176
-164: 48 176
-169: 153 180
-170: 172 181
-165: 74 182
-173: 246 184
-166: 110 186
-188: 222 190
-177: 16 191
-159: 187 192
-187: 200 192
-179: 69 194
-178: 41 195
-176: 5 196
-167: 130 196
-181: 100 201
-183: 120 202
-168: 138 202
-171: 257 202
-180: 83 203
-186: 181 207
-182: 110 208
-184: 146 208
-199: 226 210
-191: 59 211
-185: 166 215
-189: 20 220
-196: 130 221
-198: 189 222
-193: 90 228
-197: 136 232
-194: 99 236
-195: 117 236
-190: 35 238
-192: 54 248
-
