--- src/main/scala/de/htwg/se/scotlandyard/model/coreComponent/gameInitializerComponent/stationInitializerComponent/StationInitializerInterface.scala
@@ -1,7 +1,0 @@
-package de.htwg.se.scotlandyard.model.coreComponent.gameInitializerComponent.stationInitializerComponent
-
-import de.htwg.se.scotlandyard.model.Station
-
-trait StationInitializerInterface {
-  def initStations(): List[Station]
-}
