+++ src/test/scala/de/htwg/se/scotlandyard/controller/ControllerSpec.scala
@@ -4,7 +4,6 @@ import de.htwg.se.scotlandyard.controllerComponent.controllerBaseImpl.Controller
 import de.htwg.se.scotlandyard.model.{Station, StationType, TicketType}
 import de.htwg.se.scotlandyard.model.coreComponent.GameMaster
 import de.htwg.se.scotlandyard.model.coreComponent.gameInitializerComponent.gameInitializerMockImpl.GameInitializer
-import de.htwg.se.scotlandyard.model.coreComponent.gameInitializerComponent.stationInitializerComponent.stationInitializerMockImpl.StationInitializer
 import de.htwg.se.scotlandyard.model.fileIOComponent.fileIOMockImpl.FileIO
 import org.scalatest._
 
@@ -13,9 +12,6 @@ class ControllerSpec extends WordSpec with Matchers with PrivateMethodTester {
     "new" should {
       val gameInitializer = new GameInitializer()
       gameInitializer.initialize(5)
-
-      val stationInitializer = new StationInitializer()
-      stationInitializer.initStations()
 
       val controller = new Controller(gameInitializer, new FileIO(gameInitializer))
